Restructure ts-performance test with strict baseline and CI step
