+++ .circleci/config.yml
@@ -464,8 +464,8 @@ jobs:
           name: test exported typescript
           command: pnpm --filter @mui-internal/test-module-resolution typescript:all
       - run:
-          name: Run tests that depend on built packages
-          command: pnpm -r run release:test
+          name: Test theme type instantiations
+          command: pnpm --filter @mui-internal/test-ts-performance test:performance
       - store_artifacts:
           name: Persist size snapshot as pipeline artifact
           path: size-snapshot.json
