+++ test/ts-performance/README.md
@@ -1,17 +1,23 @@
-# ts-performance
+# TypeScript theme performance tests
 
-Guards the TypeScript instantiation count of the `optimizedTheme` type feature
-(`TypeFeatures` augmentation in `@mui/material/styles`).
+This suite uses workspace linking to resolve build folders and guards the
+`optimizedTheme` type feature (`TypeFeatures` augmentation in `@mui/material/styles`).
 
-The program in `src/` enables the flag, creates a theme that customizes every component,
-imports from the root barrel, and selectively re-enables type safety for `MuiButton`.
-Without the flag the same program produces >100k type instantiations; with the flag
-it must stay below the limit defined in `instantiations.test.mjs`.
+`theme.tsx` is a shared fixture that creates a theme customizing every component.
+Two TypeScript projects check it:
 
-The test resolves `@mui/material` to the build output, so it must run after
-`pnpm release:build` (wired in CI as `pnpm -r run release:test`). Run locally with:
+- `tsconfig.strict.json` (`strict/`) does not enable the flag. It asserts with
+  `@ts-expect-error` that `theme.components` keeps the strict types by default,
+  and reports the baseline instantiation count (>100k).
+- `tsconfig.optimized.json` (`optimized/`) enables the flag, imports from the root
+  barrel, and selectively re-enables `MuiButton` type safety. The instantiation
+  count must stay below the limit in `test.mjs`, and an `@ts-expect-error` fixture
+  asserts that the selective augmentation still rejects invalid values.
+
+You can only run `pnpm test:performance` after the project has been built with
+`pnpm release:build` (wired in CI after the build step). Run locally with:
 
 ```bash
-pnpm --filter @mui/material... build
-pnpm run release:test
+pnpm --filter "@mui/material..." build
+pnpm --filter @mui-internal/test-ts-performance test:performance
 ```
