+++ test/ts-performance/optimized/barrel.ts
