+++ test/ts-performance/tsconfig.optimized.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "."
+  },
+  "include": ["theme.tsx", "optimized"]
+}
