+++ test/ts-performance/tsconfig.strict.json
@@ -1,0 +1,7 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "rootDir": "."
+  },
+  "include": ["theme.tsx", "strict"]
+}
